chore: update project configuration and dependencies
- Added new entries to .gitignore for .kilo and .playwright-mcp directories to prevent unnecessary files from being tracked.
- Updated frontend package.json to include new dependencies: @react-three/fiber, chroma-js, class-variance-authority, lucide-react, motion, and @types/three, enhancing the project's capabilities.
- Removed obsolete dashboard-rail components and tests to streamline the codebase and improve maintainability.
- Introduced GraphRoute in routeTree.gen.ts and updated related components to support the new graph feature in the terminal interface.

+++ signal/pumpdump/signal_test.go
@@ -126,7 +126,7 @@ func TestMeasure(t *testing.T) {
 			thesis := wired.Thesis
 			So(market.Warmup(tests.Idle), ShouldBeNil)
 			measurements := []*types.Measurement{}
-			maturity := make(map[string]float64, len(market.Symbols))
+			maturity := make(map[string]int64, len(market.Symbols))
 
 			thesis.Measurements.Range(func(_, value any) bool {
 				for _, measurement := range value.([]*types.Measurement) {
